[Refactor] Clean up scan connector module dependencies
