+++ be/src/exec/dict_decode_node.h
@@ -22,8 +22,8 @@
 #include "common/global_types.h"
 #include "common/statusor.h"
 #include "compute_env/global_dict/parser.h"
-#include "exec/olap_common.h"
 #include "exec/pipeline_node.h"
+#include "storage_primitive/olap_scan_keys.h"
 #include "types/type_descriptor.h"
 
 namespace starrocks {
