+++ be/src/exec/intersect_node.h
@@ -25,10 +25,10 @@
 #include "column/runtime_type_traits.h"
 #include "common/statusor.h"
 #include "exec/intersect_hash_set.h"
-#include "exec/olap_common.h"
 #include "exec/pipeline_node.h"
 #include "exprs/expr_context.h"
 #include "runtime/mem_pool.h"
+#include "storage_primitive/olap_scan_keys.h"
 
 namespace starrocks {
 class DescriptorTbl;
