+++ be/src/exec/pipeline/scan/glm_manager.h
@@ -36,27 +36,6 @@ using RowsetSharedPtr = std::shared_ptr<Rowset>;
 
 namespace starrocks {
 
-class IcebergGlobalLateMaterilizationContext : public GlobalLateMaterilizationContext {
-public:
-    int32_t assign_scan_range_id(const THdfsScanRange& scan_range) {
-        std::unique_lock lock(_mutex);
-        hdfs_scan_ranges.push_back(scan_range);
-        return hdfs_scan_ranges.size() - 1;
-    }
-
-    const THdfsScanRange& get_hdfs_scan_range(int32_t scan_range_id) const {
-        std::shared_lock lock(_mutex);
-        DCHECK(scan_range_id < hdfs_scan_ranges.size())
-                << "scan_range_id: " << scan_range_id << ", size: " << hdfs_scan_ranges.size();
-        return hdfs_scan_ranges[scan_range_id];
-    }
-
-    mutable std::shared_mutex _mutex;
-    std::vector<THdfsScanRange> hdfs_scan_ranges;
-    THdfsScanNode hdfs_scan_node;
-    TPlanNode plan_node;
-};
-
 class OlapScanLazyMaterializationContext : public GlobalLateMaterilizationContext {
 public:
     void capture_rowsets(int32_t tablet_id, int64_t version, const std::vector<RowsetSharedPtr>& rowsets);
