+++ be/src/exec/pipeline/scan/olap_chunk_source.h
@@ -18,17 +18,17 @@
 
 #include "common/runtime_profile.h"
 #include "compute_env/workgroup/work_group_fwd.h"
-#include "exec/olap_common.h"
 #include "exec/olap_scan_prepare.h"
 #include "exec/pipeline/scan/chunk_source.h"
 #include "exprs/expr.h"
 #include "exprs/expr_context.h"
 #include "gen_cpp/InternalService_types.h"
 #include "runtime/runtime_state_fwd.h"
-#include "storage/query/olap_scan_range.h"
 #include "storage/tablet.h"
 #include "storage/tablet_reader.h"
 #include "storage_primitive/conjunctive_predicates.h"
+#include "storage_primitive/olap_scan_keys.h"
+#include "storage_primitive/olap_scan_range.h"
 #include "storage_primitive/predicate_tree/predicate_tree.hpp"
 
 namespace starrocks {
