+++ be/src/storage/tablet_reader_params.h
@@ -23,10 +23,10 @@
 #include "compute_env/runtime_range_pruner.h"
 #include "options.h"
 #include "storage/olap_common.h"
-#include "storage/runtime_filter_predicate.h"
 #include "storage_primitive/chunk_iterator.h"
 #include "storage_primitive/olap_tuple.h"
 #include "storage_primitive/predicate_tree/predicate_tree.hpp"
+#include "storage_primitive/runtime_filter_predicate.h"
 
 namespace starrocks {
